Move component handler tests into component folders.
