+++ v3/src/components/graph/component-handler-graph.test.ts
@@ -1,15 +1,15 @@
 import { getSnapshot } from "mobx-state-tree"
-import { INumericAxisModel } from "../../components/axis/models/axis-model"
-import { kGraphIdPrefix } from "../../components/graph/graph-defs"
-import "../../components/graph/graph-registration"
-import { IGraphContentModel, isGraphContentModel } from "../../components/graph/models/graph-content-model"
+import { V2GetGraph } from "../../data-interactive/data-interactive-component-types"
+import { DIComponentInfo } from "../../data-interactive/data-interactive-types"
+import { diComponentHandler } from "../../data-interactive/handlers/component-handler"
+import { setupTestDataset, testCases } from "../../data-interactive/handlers/handler-test-utils"
+import { testGetComponent } from "../../data-interactive/handlers/component-handler-test-utils"
 import { appState } from "../../models/app-state"
 import { toV3Id } from "../../utilities/codap-utils"
-import { V2GetGraph } from "../data-interactive-component-types"
-import { DIComponentInfo } from "../data-interactive-types"
-import { diComponentHandler } from "./component-handler"
-import { setupTestDataset, testCases } from "./handler-test-utils"
-import { testGetComponent } from "./component-handler-test-utils"
+import { INumericAxisModel } from "../axis/models/axis-model"
+import { kGraphIdPrefix } from "./graph-defs"
+import "./graph-registration"
+import { IGraphContentModel, isGraphContentModel } from "./models/graph-content-model"
 
 describe("DataInteractive ComponentHandler Graph", () => {
   const handler = diComponentHandler
